Fix a few narrowing conversions (#761)
* Fix narrowing conversion in album cover loader

* Fix narrowing conversions in discogs cover provider

* Fix narrowing conversions in spotify cover provider

* Add explicit conversion in moodbarbuilder

* Fix narrowing conversions in osd dbus

* Make WordyTimeNanosec use unsigned quint64

* Fix narrowing conversions in song

* Fix narrowing conversions in qobuz stream url request

* Make ConnectionInfo.msgLen use unsigned quint64

* Make AnalizerBase.timeout to signed int

* Fix narrowing conversions in album cover fetcher

* Make fht type be unsigned int

* Correct for type in blockanalizer and use std::fill where possible

* Revert "Fix narrowing conversions in song"

This reverts commit 3de291394df90ecb9586041709b7942f3307cfcc.

+++ src/covermanager/discogscoverprovider.h
@@ -90,8 +90,8 @@ class DiscogsCoverProvider : public JsonCoverProvider {
 
  private slots:
   void FlushRequests();
-  void HandleSearchReply(QNetworkReply *reply, const int id);
-  void HandleReleaseReply(QNetworkReply *reply, const int id, const quint64 release_id);
+  void HandleSearchReply(QNetworkReply *reply, const quint64 id);
+  void HandleReleaseReply(QNetworkReply *reply, const quint64 id, const quint64 release_id);
 
  private:
   static const char *kUrlSearch;
